+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## Change log
 ----------------------
+- v3.0.0 - Promise API & fillTopRow
 - v2.5.2 - fix stream memory limit (issue #64)
 - v2.5.1 - security fixes (npm audit)
-- v2.5.0 - Fixed nester array issue
+- v2.5.0 - fix nested array issue
 - v2.3.0 - added mapHeaders option
   - ran npm audit fix
 - v2.2.0 - Better browser distribution